+++ qualang_tools/voltage_gates/voltage_gate_sequence.py
@@ -25,7 +25,6 @@ class VoltageGateSequence:
         self._elements = elements
         # The OPX configuration
         self._config = configuration
-        self._check_OPX1000()
         # Initialize the current voltage level for sticky elements
         self.current_level = [0.0 for _ in self._elements]
         # Relevant voltage points in the charge stability diagram
@@ -34,28 +33,18 @@ class VoltageGateSequence:
         self.average_power = [0 for _ in self._elements]
         self._expression = None
         self._expression2 = None
+        self.base_operation = {}
         # Add to the config the step operation (length=16ns & amp=0.25V)
         for el in self._elements:
-            self._config["elements"][el]["operations"]["step"] = "step_pulse"
-        self._config["pulses"]["step_pulse"] = {
-            "operation": "control",
-            "length": 16,
-            "waveforms": {"single": "step_wf"},
-        }
-        self._config["waveforms"]["step_wf"] = {"type": "constant", "sample": 0.25}
-
-    def _check_OPX1000(self):
-        opx1000 = False
-
-        for con in self._config["controllers"].keys():
-            if "type" in self._config["controllers"][con].keys():
-                if self._config["controllers"][con]["type"] == "opx1000":
-                    opx1000 = True
-            if opx1000:
-                warn(
-                    "A bug currently prevents the VoltageGateSequence from working when using the amplified mode of the OPX1000 LF-FEM with ramps.",
-                    stacklevel=2,
-                )
+            base_amplitude = self._check_amplified_mode(el)
+            self.base_operation[el] = {"amplitude": base_amplitude, "bit_shift": int(np.log2(1 / base_amplitude))}
+            self._config["elements"][el]["operations"]["step"] = f"{el}_step_pulse"
+            self._config["pulses"][f"{el}_step_pulse"] = {
+                "operation": "control",
+                "length": 16,
+                "waveforms": {"single": f"{el}_step_wf"},
+            }
+            self._config["waveforms"][f"{el}_step_wf"] = {"type": "constant", "sample": base_amplitude}
 
     def _check_amplified_mode(self, element: str):
         con = self._config["elements"][element]["singleInput"]["port"][0]
@@ -68,9 +57,9 @@ class VoltageGateSequence:
                         self._config["controllers"][con]["fems"][fem]["analog_outputs"][ch]["output_mode"]
                         == "amplified"
                     ):
-                        raise RuntimeWarning(
-                            "A bug currently prevents the VoltageGateSequence from working when using the amplified mode of the OPX1000 LF-FEM with ramps."
-                        )
+                        return 0.5
+
+        return 0.25
 
     def _check_name(self, name, key):
         if name in key:
@@ -204,7 +193,11 @@ class VoltageGateSequence:
                 if self.is_QUA(voltage_level) or self.is_QUA(self.current_level[i]):
                     # if dynamic duration --> play step and wait
                     if self.is_QUA(_duration):
-                        play("step" * amp((voltage_level - self.current_level[i]) * 4), gate)
+                        play(
+                            "step"
+                            * amp((voltage_level - self.current_level[i]) << self.base_operation[gate]["bit_shift"]),
+                            gate,
+                        )
                         wait((_duration - 16) >> 2, gate)
                     # if constant duration --> new operation and play(*amp(..))
                     else:
@@ -212,10 +205,16 @@ class VoltageGateSequence:
                             operation = self._add_op_to_config(
                                 gate,
                                 "step",
-                                amplitude=0.25,
+                                amplitude=self.base_operation[gate]["amplitude"],
                                 length=_duration,
                             )
-                            play(operation * amp((voltage_level - self.current_level[i]) * 4), gate)
+                            play(
+                                operation
+                                * amp(
+                                    (voltage_level - self.current_level[i]) << self.base_operation[gate]["bit_shift"]
+                                ),
+                                gate,
+                            )
 
                 # Fixed amplitude but dynamic duration --> new operation and play(duration=..)
                 elif isinstance(_duration, (QuaExpression, QuaVariable)):
@@ -284,9 +283,11 @@ class VoltageGateSequence:
                     duration_4ns = max((int(np.ceil(comp_duration)) // 4 + 1) * 4, 48)
                     # Corrected amplitude to account for the duration casting to integer
                     amplitude = -np.sign(self.average_power[i]) * max_amplitude * comp_duration / duration_4ns
-                    # 3 cc gap between compensation pulse and ramp to zero
-                    operation = self._add_op_to_config(gate, "compensation", amplitude=0.25, length=duration_4ns - 12)
-                    play(operation * amp((amplitude - self.current_level[i]) * 4), gate)
+                    # Apply the compensation pulse as a ramp to circumvent the max amplitude limit.
+                    ramp_rate = (amplitude - self.current_level[i]) / 16
+                    play(ramp(ramp_rate), gate, duration=4)
+                    wait((duration_4ns - 16) // 4, gate)
+
                 else:
                     amplitude = -0.0009765625 * self.average_power[i] / duration
                     operation = self._add_op_to_config(
@@ -295,7 +296,6 @@ class VoltageGateSequence:
                     play(operation, gate)
             else:
                 if duration is None:
-                    operation = self._add_op_to_config(gate, "compensation", amplitude=0.25, length=16)
                     eval_average_power = declare(int)
                     comp_duration = declare(int)
                     duration_4ns = declare(int)
@@ -303,10 +303,10 @@ class VoltageGateSequence:
                     duration_4ns_pow2_cur = declare(int)
                     amplitude = declare(fixed)
                     # Exact duration of the compensation pulse
-                    # take into account a gap of 96ns for the derivation of the compensation pulse
+                    # take into account a gap of 110ns for the derivation of the compensation pulse
                     assign(
                         eval_average_power,
-                        self.average_power[i] + Cast.mul_int_by_fixed(256 * 1024, self.current_level[i]),
+                        self.average_power[i] + Cast.mul_int_by_fixed(96 * 1024, self.current_level[i]),
                     )
                     assign(
                         comp_duration, Cast.mul_int_by_fixed(Math.abs(eval_average_power), 0.0009765625 / max_amplitude)
@@ -325,13 +325,16 @@ class VoltageGateSequence:
                         assign(amplitude, -Cast.mul_fixed_by_int(max_amplitude >> duration_4ns_pow2, comp_duration))
                     with else_():
                         assign(amplitude, Cast.mul_fixed_by_int(max_amplitude >> duration_4ns_pow2, comp_duration))
-                    play(
-                        operation * amp((amplitude - self.current_level[i]) << 2),
-                        gate,
-                        duration=duration_4ns_pow2_cur >> 2,
-                    )
+
+                    # Apply the compensation pulse as a ramp to circumvent the max amplitude limit.
+                    ramp_rate = declare(fixed)
+                    assign(ramp_rate, (amplitude - self.current_level[i]) >> 4)
+                    play(ramp(ramp_rate), gate, duration=4)
+                    wait((duration_4ns_pow2_cur - 16) >> 2, gate)
                 else:
-                    operation = self._add_op_to_config(gate, "compensation", amplitude=0.25, length=duration)
+                    operation = self._add_op_to_config(
+                        gate, "compensation", amplitude=self.base_operation[gate]["amplitude"], length=duration
+                    )
                     amplitude = declare(fixed)
                     eval_average_power = declare(int)
                     assign(eval_average_power, self.average_power[i])
@@ -339,7 +342,10 @@ class VoltageGateSequence:
                     # The calculation is done in two steps to avoid rounding errors
                     assign(amplitude, -Cast.mul_fixed_by_int(0.01 / duration, eval_average_power))
                     assign(amplitude, amplitude * 0.09765625)
-                    play(operation * amp((amplitude - self.current_level[i]) * 4), gate)
+                    play(
+                        operation * amp((amplitude - self.current_level[i]) << self.base_operation[gate]["bit_shift"]),
+                        gate,
+                    )
 
             self.current_level[i] = amplitude